Redirect to show page after edit
