+++ client/src/semantic-data-provider/dataProvider.js
@@ -1,4 +1,5 @@
 import jsonld from 'jsonld';
+import buildSparqlQuery from "./buildSparqlQuery";
 
 const getJsonContext = (ontologies, mainOntology) => {
   let pattern = {};
@@ -9,49 +10,6 @@ const getJsonContext = (ontologies, mainOntology) => {
   } else {
     return pattern;
   }
-};
-
-const getPrefixRdf = ontologies => {
-  return ontologies.map(ontology => `PREFIX ${ontology.prefix}: <${ontology.url}>`).join('\n');
-};
-
-const computeSparqlQuery = ({ types, params: { query, pagination, sort, filter }, ontologies }) => {
-  let whereQuery = '';
-
-  if (filter.q && filter.q.length > 0) {
-    whereQuery += `
-      {
-        SELECT ?s1
-        WHERE {
-          ?s1 ?p1 ?o1 .
-          FILTER regex(lcase(str(?o1)), "${filter.q.toLowerCase()}")
-          FILTER NOT EXISTS {?s1 a ?o1}
-        }
-      }
-      `;
-  }
-  if (query) {
-    Object.keys(query).forEach(predicate => {
-      const value = query[predicate].startsWith('http') ? `<${query[predicate]}>` : query[predicate];
-      whereQuery += `?s1 ${predicate} ${value} .`;
-    });
-  }
-  return `
-    ${getPrefixRdf(ontologies)}
-    CONSTRUCT {
-      ?s1 ?p2 ?o2
-    }
-    WHERE {
-      ${whereQuery}
-      ?s1 a ?type .
-      FILTER( ?type IN (${types.join(', ')}) ) .
-      FILTER( (isIRI(?s1)) ) .
-      ?s1 ?p2 ?o2 .
-    }
-    # TODO try to make pagination work in SPARQL as this doesn't work.
-    # LIMIT ${pagination.perPage}
-    # OFFSET ${(pagination.page - 1) * pagination.perPage}
-  `;
 };
 
 const dataProvider = ({ sparqlEndpoint, httpClient, resources, ontologies, mainOntology }) => ({
@@ -86,7 +44,7 @@ const dataProvider = ({ sparqlEndpoint, httpClient, resources, ontologies, mainO
       /*
        * Do a SPARQL search
        */
-      const sparqlQuery = computeSparqlQuery({
+      const sparqlQuery = buildSparqlQuery({
         types: resources[resourceId].types,
         params: { ...params, query: resources[resourceId].query },
         ontologies
@@ -160,11 +118,10 @@ const dataProvider = ({ sparqlEndpoint, httpClient, resources, ontologies, mainO
     const headers = new Headers();
 
     if( slugField ) {
-      if( Array.isArray(slugField) ) {
-        headers.set('Slug', slugField.map(f => params.data[f]).join(' '));
-      } else {
-        headers.set('Slug', params.data[slugField]);
-      }
+      headers.set('Slug', Array.isArray(slugField)
+        ? slugField.map(f => params.data[f]).join(' ')
+        : params.data[slugField]
+      );
     }
 
     const { headers: responseHeaders } = await httpClient(containerUri, {
